fixes made istio access log test work
