+++ tests/fast-integration/logging/loki.js
@@ -14,7 +14,10 @@ const {
   logsPresentInLoki,
   queryLoki,
 } = require('./client');
-const {info} = require('../utils');
+const {
+  info,
+  sleep,
+} = require('../utils');
 
 async function checkCommerceMockLogs(startTimestamp) {
   const labels = '{app="commerce-mock", container="mock", namespace="mocks"}';
@@ -59,6 +62,9 @@ function parseJson(str) {
 
 async function verifyIstioAccessLogFormat(startTimestamp) {
   console.log('timestamp: '+ startTimestamp + '\n');
+  await sleep(10*1000);
+  timeNow = new Date().toISOString();
+  console.log('time now: '+ timeNow + '\n');
   const query = '{container="istio-proxy",namespace="kyma-system",pod="logging-loki-0"}';
   const responseBody = await queryLoki(query, startTimestamp);
   assert.isTrue(responseBody.data.result[0].values.length > 0, 'No Istio access logs found for loki');
@@ -115,6 +121,9 @@ async function verifyIstioAccessLogFormat(startTimestamp) {
       return;
     }
   }
+
+  timeNow1 = new Date().toISOString();
+  console.log('time at end: '+ timeNow1 + '\n');
   log = parseJson(entry.log);
   assert.isDefined(log, `Istio access log is not in JSON format: ${entry.log}`);
 }
