+++ tests/fast-integration/test/test.js
@@ -6,9 +6,10 @@ const {
 const {apiExposureTests} = require('../api-exposure');
 const {monitoringTests, unexposeGrafana} = require('../monitoring');
 const {loggingTests} = require('../logging');
+const {createIstioAccessLogResource} = require('../logging/client.js');
 const {cleanMockTestFixture} = require('./fixtures/commerce-mock');
 const {ensureCommerceMockLocalTestFixture} = require('../test/fixtures/commerce-mock');
-const {tracingTests} = require('../tracing');
+// const {tracingTests} = require('../tracing');
 const {error} = require('../utils');
 
 describe('Executing Standard Testsuite:', function() {
@@ -23,6 +24,7 @@ describe('Executing Standard Testsuite:', function() {
       error(err);
       return ensureCommerceMockLocalTestFixture(mockNamespace, testNamespace);
     });
+    await createIstioAccessLogResource();
   });
 
   after('Test Cleanup: Test namespaces should be deleted', async function() {
@@ -41,5 +43,5 @@ describe('Executing Standard Testsuite:', function() {
   // gettingStartedGuideTests();
 
   loggingTests();
-  tracingTests(testNamespace);
+  // tracingTests(testNamespace);
 });
